+++ spec/graphql/fragment_cache/rails/cache_key_builder_spec.rb
@@ -39,7 +39,7 @@ describe GraphQL::FragmentCache::CacheKeyBuilder do
   it "uses Cache.expand_cache_key" do
     allow(ActiveSupport::Cache).to receive(:expand_cache_key).with(object) { "as:cache:key" }
 
-    is_expected.to eq "schema_key/cachedPost(id:#{id})[id.title.author[id.name]]/Post#42"
+    is_expected.to eq "graphql/schema_key/cachedPost(id:#{id})[id.title.author[id.name]]/Post#42"
   end
 
   context "when object is passed and responds to #graphql_cache_key" do
@@ -47,7 +47,7 @@ describe GraphQL::FragmentCache::CacheKeyBuilder do
       object.singleton_class.define_method(:graphql_cache_key) { "{graphql_cache_key}" }
     end
 
-    specify { is_expected.to eq "schema_key/cachedPost(id:#{id})[id.title.author[id.name]]/{graphql_cache_key}" }
+    specify { is_expected.to eq "graphql/schema_key/cachedPost(id:#{id})[id.title.author[id.name]]/{graphql_cache_key}" }
   end
 
   context "when object is passed and responds to #cache_key_with_version" do
@@ -55,7 +55,7 @@ describe GraphQL::FragmentCache::CacheKeyBuilder do
       object.singleton_class.define_method(:cache_key_with_version) { "{cache_key_with_version}" }
     end
 
-    specify { is_expected.to eq "schema_key/cachedPost(id:#{id})[id.title.author[id.name]]/{cache_key_with_version}" }
+    specify { is_expected.to eq "graphql/schema_key/cachedPost(id:#{id})[id.title.author[id.name]]/{cache_key_with_version}" }
   end
 
   context "when object is passed and responds to #cache_key" do
@@ -63,6 +63,6 @@ describe GraphQL::FragmentCache::CacheKeyBuilder do
       object.singleton_class.define_method(:cache_key) { "{cache-key}" }
     end
 
-    specify { is_expected.to eq "schema_key/cachedPost(id:#{id})[id.title.author[id.name]]/{cache-key}" }
+    specify { is_expected.to eq "graphql/schema_key/cachedPost(id:#{id})[id.title.author[id.name]]/{cache-key}" }
   end
 end